fix C++17 build
As I understood it, C++17 allows [aggregate initialisation][1] for classes
which have public base classes. This changes the line from a statement
calling the default constructor, to a statement performing aggregate
initialisation, and the base class is default initialized.

In other words, in C++17, the line:

return RequestBuilder{};

is understood as:

return RequestBuilder{CommonBuilder<...>{}};

And since the default-constructor of CommonBuilder is protected, GCC and
Clang errors out.

Here, we explicitely call the default constructor, implicitely defined
in RequestBuilder.

[1]: https://en.cppreference.com/w/cpp/language/aggregate_initialization